% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check.genotype.probs.r
\name{check.genotype.probs}
\alias{check.genotype.probs}
\title{Check \code{genotype.probs} object}
\usage{
check.genotype.probs(geno, lmm.l1o = FALSE)
}
\arguments{
\item{geno}{a \code{genotype.probs} object}
}
\value{
\code{NULL}
}
\description{
Check \code{genotype.probs} object
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

geno <- extract.geno(fake.f2)
HPQTL:::check.genotype.probs
}
\keyword{manip}

