% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mvnpermute}
\alias{mvnpermute}
\title{Permuted Y for (QTLRel) Calculation of Permutation Threshold}
\usage{
mvnpermute(y, X, S, nr)
}
\arguments{
\item{y}{vector of length n containing the observed trait data}

\item{X}{n x m matrix or data frame, in which each row corresponds to a
       sample and each column corresponds to a covariate. If the
       model includes an intercept term, this should be included as
       one of the covariates of this matrix, as in

        \code{X <- cbind(rep(n,1),X)}}

\item{S}{known or estimated covariance matrix of the data.}

\item{nr}{number of replicates to generate.}
}
\value{
The function returns a matrix with n rows and nr columns, in
which each column is a "permuted" version of y.
}
\description{
R function "mvnpermute" for executing a permutation-based test with
multivariate normally distributed data.
}
\details{
This function takes multivariate normal data with known covariates
and covariance matrix and generates "permutations" of this data that
maintain the mean and covariance of the original data. The
permutations are generated by finding the residuals of the data and
mapping the residuals to an orthonormal space, then simulating
random permutations within this orthonormal space. The permutations
are then mapped back to the original space, and the expected values
(given the covariates, and their estimated effect sizes) are added
back in. The resulting "permuted" data can now be used exactly like
the original data; for example, you could estimate the null
distribution of some test statistic while observing the (known)
covariance of the data.
}
\author{
Mark Abney
}
\references{
Abney M, Ober C, McPeek MS (2002). "Quantitative trait homozygosity
and association mapping and empirical genome-wide significance in
large complex pedigrees: Fasting serum insulin levels in the
Hutterites." American Journal of Human Genetics 70: 920-934.
}
\keyword{manip}

