% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize.matrix}
\alias{normalize.matrix}
\title{Normalize Matrix}
\usage{
normalize.matrix(W, method = c("sample-variance", "diagonal",
  "diagonal-average", "none"))
}
\arguments{
\item{W}{symmetric, positive semidefinite matrix}

\item{method}{"sample-variance", "diagonal" or "diagonal-average"}
}
\value{
Normalized matrix
}
\description{
Normalize genetic similarity matrix.
}
\details{
To calculate heritability, particularly for different genetic similarity estimators,
the genetic similarity matrix needs to be properly normalized. The default way is to force sample
variance of animal effect to be unit.
}
\examples{
W <- matrix(c(1,1/2,1/2,1),2,2)
normalize.matrix(W, method="sample-variance")
normalize.matrix(W, method="diagonal")
}
\keyword{manip}

