% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankZ.r
\name{rankZ}
\alias{rankZ}
\title{RankZ Normalization transform}
\usage{
rankZ(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Normalized vector.
}
\description{
Return quantiles normalized (rank-Z) transformed vector.
}
\examples{
x <- 2 ^ rnorm(1000)
hist(x)

x_rankz <- rankZ(x)
hist(x_rankz)
}
\keyword{manip}

