% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan1.r
\name{scan1}
\alias{scan1}
\title{Genome Scan With a Single QTL Model}
\usage{
scan1(geno, pheno, pheno.col = 1, rankZ = FALSE, covar = NULL,
  intcovar = NULL, procedure = c("LM", "LMM", "LOCO"), G, verbose = FALSE,
  Intercept = rep(1, length(geno$subjects)), ...)
}
\arguments{
\item{geno}{genotype probabilities ("\code{genotype.probs}" or "\code{cross}" object)}

\item{pheno}{data frame with phenotypes}

\item{covar}{(additive) covariates}

\item{intcovar}{interactive covariates (interact with QTL genotype)}

\item{procedure}{procedure to do the inference, see Details}

\item{G}{genetic similarity matrix or a list of genetic similarity matrices or \code{NULL}}

\item{Intercept}{option to pass rotated intercept for LMM model}

\item{...}{parameters passed to \code{gensim.matrix}}

\item{pheno.cols}{selection of phenotype's column(s)}
}
\value{
\code{scanone} object
}
\description{
Genome Scan With a Single QTL Model
}
\details{
Currently, three procedures are implemented: linear model (LM), linear mixed model (LMM)
and linear mixed model - leave one chromosome out (LOCO).
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

plot(scan1(fake.f2, procedure="LM"), incl.markers=FALSE)
plot(scan1(fake.f2, procedure="LMM"), incl.markers=FALSE)
plot(scan1(fake.f2, procedure="LOCO"), incl.markers=FALSE)
}
\keyword{manip}

