% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.cross.geno.r
\name{sim.cross.geno}
\alias{sim.cross.geno}
\title{Generate a Genotype of Random Cross}
\usage{
sim.cross.geno(nind = 250, ctype = c("f2", "bc"), nchr = 20,
  chrlen = 100, nmar = 10)
}
\arguments{
\item{nind}{number of animals}

\item{ctype}{cross type ("f2" / "bc")}

\item{nchr}{number of chromosomes}

\item{chrlen}{length of chromosomes (in cM)}

\item{nmar}{number of markers per chromosome}
}
\value{
"\code{cross}" object
}
\description{
Generate a Genotype of Random Cross
}
\examples{
sim.cross.geno(250, nmar=10)
}
\keyword{manip}

