% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.qtl.r
\name{add.qtl}
\alias{add.qtl}
\title{Add a QTL to Simulated Phenotype}
\usage{
add.qtl(geno, pheno, qtl.pos, qtl.size, pheno.cols = 1)
}
\arguments{
\item{qtl.pos}{position of QTL(s), either number of markers or relative position in genome (between 0 and 1)}

\item{qtl.size}{size of QTL(s),}

\item{pheno.cols}{selection of phenotype's column(s)}

\item{cross}{"\code{cross}" object}
}
\value{
"\code{cross}" object
}
\description{
Add a QTL to Simulated Phenotype
}
\examples{
geno <- sim.cross.geno(250, nmar=10)
pheno <- sim.cross.pheno(0.01, geno)
plot(scan1(geno, pheno))
pheno2 <- add.qtl(geno, pheno, 0.75, 1)
plot(scan1(geno, pheno2))
}
\keyword{manip}

