% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.geno}
\alias{extract.geno}
\title{Import qtl / DOQTL cross}
\usage{
extract.geno(cross)
}
\arguments{
\item{cross}{"\code{cross}" object or DO qtl}
}
\value{
\code{genotype.probs} object is a list of three components
\itemize{
  \item \code{probs} 3-dimensional array of genotype probabilities (subjects x calls x markers) or a list of such arrays
  \item \code{calls} possible genotype calls
  \item \code{markers}
}

\code{geno} object
}
\description{
Import qtl / DOQTL cross
}
\details{
Extract genotype prob. array from "\code{qtl::cross}" object.
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)
geno <- extract.geno(fake.f2)
}
\keyword{manip}

