% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gensim.matrix.r
\name{gensim.matrix}
\alias{gensim.matrix}
\title{Genetic Similarity Matrix}
\usage{
gensim.matrix(geno, procedure = c("LMM", "LOCO", "POMOLOCO", "LM"),
  gensim.normalization = "sample-variance", verbose = FALSE)
}
\arguments{
\item{geno}{genotype probabilities ("\code{genotype.probs}" or "\code{cross}" object)}

\item{procedure}{procedure of scan1 the G is calculated for}

\item{gensim.normalization}{method to be used for genetic similarity matrix normalization}

\item{subjects}{subseting of subjects}

\item{markers}{subseting of markers}
}
\value{
square matrix (\code{nind} rows, \code{nind} columns)
}
\description{
Genetic Similarity Matrix
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

geno <- extract.geno(fake.f2)
G <- gensim.matrix(geno)
Glist <- gensim.matrix(geno, procedure="LOCO")
}
\keyword{manip}

