% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gensim.matrix}
\alias{gensim.matrix}
\title{Genetic Similarity Matrix}
\usage{
gensim.matrix(geno, method = c("default", "allele-2f-additive",
  "allele-multif-additive", "allele-2f-weighted", "allele-multif-cosine"),
  procedure = c("LMM", "LOCO", "LM"), subjects = seq(geno$subjects),
  markers = seq(NROW(geno$markers)),
  gensim.normalization = "sample-variance", ...)
}
\arguments{
\item{geno}{genotype probabilities ("\code{genotype.probs}" or "\code{cross}" object)}

\item{method}{how to calculate similarity matrix}

\item{procedure}{procedure of scan1 the G is calculated for}

\item{subjects}{subseting of subjects}

\item{markers}{subseting of markers}

\item{gensim.normalization}{method to be used for genetic similarity matrix normalization}
}
\value{
square matrix (\code{nind} rows, \code{nind} columns)
}
\description{
Genetic Similarity Matrix
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

geno <- extract.geno(fake.f2)
G <- gensim.matrix(geno)
Glist <- gensim.matrix(geno, procedure="LOCO")
}
\keyword{manip}

