% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gensim.matrix}
\alias{gensim.matrix}
\title{Genetic Similarity Matrix}
\usage{
gensim.matrix(geno, method = c("default", "allele-2f-additive",
  "allele-multif-additive", "allele-multif-cosine"), procedure = c("LMM",
  "LMM-L1O", "LM"), subjects = seq(geno$subjects),
  markers = seq(NROW(geno$markers)), ...)
}
\arguments{
\item{geno}{genotype probabilities ("\code{genotype.probs}" or "\code{cross}" object)}

\item{method}{how to calculate similarity matrix}

\item{procedure}{procedure of scan1 the G is calculated for}

\item{subjects}{subseting of subjects}

\item{markers}{subseting of markers}
}
\value{
square matrix (\code{nind} rows, \code{nind} columns)
}
\description{
Genetic Similarity Matrix
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

geno <- extract.geno(fake.f2)
G <- gensim.matrix(geno)
Glist <- gensim.matrix(geno, procedure="LMM-L1O")
}
\keyword{manip}

