% Generated by roxygen2 (4.0.2): do not edit by hand
\name{half.inv}
\alias{half.inv}
\title{Half-inverse matrix operation}
\usage{
half.inv(W, symmetric = TRUE, inverse = TRUE)
}
\arguments{
\item{W}{positive definite matrix}
}
\value{
Matrix \code{A} such that \code{t(A)\%*\%A} is the inverse of \code{W}.
}
\description{
Return square root of an inverse matrix.
}
\details{
Note \code{W} must be strictly positive definite.
For a matrix with negative eigen values the error is thrown. However, the problem
might arise for matrices with eigen values close to zero.
}
\examples{
W <- matrix(c(1,1/2,1/2,1),2,2)
A <- HPQTL:::half.inv(W)
max(abs(t(A)\%*\%A\%*\%W) - diag(2))
}
\author{
Riyan Cheng, copied from QTLRel
}
\keyword{manip}

