% Generated by roxygen2 (4.0.1): do not edit by hand
\name{heritability}
\alias{heritability}
\title{Estimate Heritability}
\usage{
heritability(y, covar, G, se = FALSE, var.decompose = FALSE, ...)
}
\arguments{
\item{y}{a vector with phenotype}

\item{covar}{a matrix with covariates, Intercept should be included}

\item{G}{genetic similarity matrix}

\item{se}{if TRUE standard error is estimated}

\item{var.decompose}{if TRUE total variability is decomposed as into three parts: fixed effects, animal effect and noise}

\item{...}{parameters passed to \code{gensim.matrix}}
}
\value{
numeric
}
\description{
Estimate Heritability
}
\keyword{manip}

