% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to.additive}
\alias{to.additive}
\title{Convert 3 calls to 2 (additive) representation}
\usage{
to.additive(geno)
}
\arguments{
\item{geno}{genotype probabilities ("\code{genotype.probs}" or "\code{cross}" object)}
}
\value{
\code{genotype.probs} object with 2 possible calls
}
\description{
Convert 3 calls to 2 (additive) representation
}
\examples{
data(fake.f2, package="qtl")
fake.f2 <- calc.genoprob(fake.f2)

geno <- extract.geno(fake.f2)
geno2 <- to.additive(geno)
}
\keyword{manip}

